#7 firebase関係の処理を別ファイルに切り出して再利用性を高めた
